Logins controller eklendi. postman eklendi. company düzenlendi. logins de çalışıyor şuan. ve genel düzenlemeler.

+++ ws/src/main/java/com/burhanmutlu/ws/service/impl/CompanyServiceImpl.java
@@ -2,9 +2,9 @@ package com.burhanmutlu.ws.service.impl;
 
 import com.burhanmutlu.ws.dto.CompanyDto;
 import com.burhanmutlu.ws.dto.req.CompanyRequest;
-import com.burhanmutlu.ws.model.Company;
+import com.burhanmutlu.ws.entity.Company;
 import com.burhanmutlu.ws.exception.CompanyNotFoundException;
-import com.burhanmutlu.ws.model.User;
+import com.burhanmutlu.ws.entity.User;
 import com.burhanmutlu.ws.repository.CompanyRepository;
 import com.burhanmutlu.ws.service.CompanyService;
 import com.burhanmutlu.ws.service.UserService;
@@ -34,12 +34,19 @@ public class CompanyServiceImpl implements CompanyService {
         List<Company> companies = companyRepository.findAllByUserId(id);
 
         List<CompanyDto> companyDtoList = new ArrayList<>();
+        boolean updatable;
         for(Company company : companies) {
+            updatable = true;
+
+            if(company.getUserId().getId() == 1) {
+                updatable = false;
+            }
             CompanyDto companyDto = CompanyDto.builder()
                     .id(company.getId())
                     .companyName(company.getCompanyName())
                     .companyLogo(company.getCompanyLogo())
                     .companyWebPage(company.getCompanyWebPage())
+                    .updatable(updatable)
                     .build();
             companyDtoList.add(companyDto);
         }
@@ -50,18 +57,23 @@ public class CompanyServiceImpl implements CompanyService {
     public CompanyDto getCompanyById(Long id) {
         Company company = companyRepository.findById(id).orElseThrow(
                 () -> { throw new CompanyNotFoundException("Company not found"); });
+        boolean updatable = true;
+        if(company.getUserId().getId() == 1) {
+            updatable = false;
+        }
 
         CompanyDto companyDto = CompanyDto.builder()
                 .id(company.getId())
                 .companyName(company.getCompanyName())
                 .companyLogo(company.getCompanyLogo())
+                .updatable(updatable)
                 .companyWebPage(company.getCompanyWebPage()).build();
 
         return companyDto;
     }
 
     @Override
-    public Company addCompanyById(CompanyRequest request, Long userId) {
+    public CompanyDto addCompanyByUserId(CompanyRequest request, Long userId) {
 
         User user = userService.getUserById(userId);
 
@@ -72,18 +84,54 @@ public class CompanyServiceImpl implements CompanyService {
                 .userId(user)
                 .build();
 
+        company = companyRepository.save(company);
 
-        return companyRepository.save(company);
+        CompanyDto companyDto = CompanyDto.builder()
+                .companyName(company.getCompanyName())
+                .companyWebPage(company.getCompanyWebPage())
+                .companyLogo(company.getCompanyLogo())
+                .id(userId)
+                .updatable(true)
+                .build();
+
+        return companyDto;
     }
 
     @Override
-    public Company updateCompany(Company updatedCompany) {
-        return null;
+    public CompanyDto updateCompany(CompanyRequest companyRequest, Long id) {
+        Company company = companyRepository.findById(id).orElseThrow(
+                () -> { throw new CompanyNotFoundException("company not found"); });
+
+        if(company.getUserId().getId() == 1) {
+            throw new RuntimeException("You do not have the authority to change");
+        }
+
+        company = Company.builder()
+                .companyName(companyRequest.getCompanyName())
+                .companyLogo(companyRequest.getCompanyLogo())
+                .companyWebPage(companyRequest.getCompanyWebPage())
+                .id(id)
+                .build();
+
+        companyRepository.save(company);
+
+        CompanyDto companyDto = CompanyDto.builder()
+                .id(company.getId())
+                .companyName(company.getCompanyName())
+                .companyLogo(company.getCompanyLogo())
+                .companyWebPage(company.getCompanyWebPage())
+                .updatable(true)
+                .build();
+
+        return companyDto;
     }
 
     @Override
     public void deleteCompany(Long id) {
-        getCompanyById(id);
+        Company company = companyRepository.findById(id).orElse(null);
+        if(company == null || company.getUserId().getId() == 1) {
+            throw new RuntimeException("company not found or You do not have the authority to change");
+        }
         companyRepository.deleteById(id);
     }
 }
